(#8010) Add basic structure for group management
This change adds a basic structure to the accounts module.  Data will live in
the accounts::data namespace by default.  The user can easily customize where
the module looks for data by setting the data_namespace class parameter.

The module SHOULD look to a global parameter to support the Dashboard being
used to set this location.

The lookupvar() function is being added to the stdlib module.