(#8010) Add create_resources function
This function provides the ability to create resources from a Hash.  It
will be included in Puppet 2.7 as part of core, so the technical debt of
shipping this function in stdlib is high.

As a compromise, I'm shipping the function with the module itself.  When
the module is targeted at Puppet 2.7 the module function may be disabled
somehow.  How, I'm not yet sure.

This is an alternative to implementing my own functions that transform a
hash into an array and a hash of hashes, feeding the two results into a
defined resource that instantiates other defined resources.