(#8010) Add basic group management using create_resources()
This change uses the create_resources() function shipped with the
accounts module to declare a number of groups defined in the hash
variable named groups_hash in the data namespace.

This function "just works" for this purpose.

In addition, the groups class is contained within the accounts composite
class using the Anchor pattern.

The decision to ship create_resources() with the accounts module is a
tricky one.  create_resources() is included in Puppet 2.7 and will cause
a conflict between that release and this module.