(#8010) Add basic user management
This change follows the same structure and design of basic group
management.  The accounts::users class requires a hash to be passed as
the users_hash parameter.

The end user provides this data in the form of two variables in a
customizable namespace:

 * users_hash
 * users_hash_default

The accounts class is responsible for grabbing these two hashes out of
the namespace, merging them together, and passing the result to the
accounts::users class.

The accounts::users class is responsible for adding resources to the
catalog based on the data provided.  It accomplishes this using the
create_resources() function back ported from 2.7.x into the module.

Reviewed-by: Nigel Kersten