WIP Remove data not directly usable by the type
This change removes the sshkey parameter.  We don't currently have a way
to filter out a hash to ensure only keys directly usable by the
create_resources() function are present.

Rather than implementing a filter_type_hash() function, I'm instead
simply specifying the data in a way that only the user type may consume.

In addition, we have no way to merge Hashes presently, so FIXME comments
have been added to the code where it would be good to add this
functionality in a future iteration.

Reviewed-by: Nigel Kersten