(#8010) Add basic user and primary group management
This change uses a Ruby DSL class to add basic user resources to the
catalog.  The $users_hash user defined datastructure is brought into the
local scope of the class block.  The validation functions are called
just as they were in the Puppet DSL class.

Next, the users hash is filtered, looking only for gid and ensure
parameters.  A new hash is constructed and fed into create_resources()
to create the primary group for each user following the similar name
convention of most modern Linux distributions.

A before relationship is established to the user resource which has yet
to be added to the catalog.

Finally, the user provided $users_hash is fed as-is into
create_resources()

This satisfies the minimum requirements of these entries in #8010

 * Enforce convention of user-named primary group

Reviewed-by: Nigel Kersten