(#8010) Fill in user defined $users_hash_default
This change provides an example how the end user of the module will
specify default data to be used for specific account resources which do
not have the corresponding resource parameters.

In addition data such as sshkey is being added back into the system
because I'm about to patch the account::users class to filter out any
keys which are invalid for the resource type being declared.

Finally, I'm slipping in a feature that allows a pattern to be used for
the home directory.  The user may specify the default home directory
should be /home/#{title} and the module will sort it out for them.