(#8010) Pull the user defined $users_hash_default
This change brings the user defined default set of parameters
for the users resource into the accounts namespace using the getvar()
function.

In addition, it makes more sense to validate the user inputs as soon
as their pulled into the module rather than leaving it up to all of
the implementation classes to validate the data.

This change moves validation up into the accounts class