(#8010) Add merging with the default hash
This change adds support for the end user to specify default parameters for all
of the user resources in a variable named $users_hash_default.

Unlike the $users_hash, this is a single level parameter/value hash.

When a parameter for the user resource exists in the default hash but not in
the specific user hash, this change will merge in the value from the default
hash.

Filtering is done such that only parameters valid for the User resource type
make it through the filter and into the create_resources() function.

In addition, all parameter values are interpolated which may be a HORRIBLE idea
to expose to the end user since we're calling eval() on user supplied data.

Note, this interpolation make break down for values that are not strings.