(#8010) Add ssh key management support
If the end user of the module defines an sshkey parameter for the user
resource, then the module will manage their ~/.ssh/authorized_keys file.

If the end user does not specify this sshkey parameter, the module will
not manage the authorized_keys file at all.

It may be more prudent to always manage the file for security and only
manage the contents if the parameter value is present.