(#8010) Only interpolate the home directory param
As suspected, trying to interpolate all values of all the parameters resulted
in an error as soon as we hit a value that is not a String.

This change changes the interpolation to only happen on the home parameter
which should always be a string.

It still may be better to use the % operator to interpolate the string than
eval.