(#8575) Add YAML data file support
This change gives the end user the ability to specify the following
three data structures as a simple YAML file:
 * users_hash
 * users_hash_default
 * groups_hash

This is in contrast to only being able to define accounts and groups
using a Puppet hash variable.

The data must be stored in Puppet $confdir/data/ using the following
three file names:
 * accounts_groups_hash.yaml
 * accounts_users_default_hash.yaml
 * accounts_users_hash.yaml

This change also makes the "yaml" data store the default for the module.
The end user may switch back to data being defined in a Puppet namespace
by using:

    class { 'accounts': data_store => namespace }

Which will default to looking for variables in accounts::data

Reviewed-by: Dan Bode