(#8577) Support multiple ssh keys per user
This change adds the ability for the end user of the module to specify
more than one SSH key in the data structure.  Before this change, only 0
or 1 key was supported.

Additional changes in behavior introduced with this patch:

 * Don't manage (clobber) the entire authorized_keys file
 * Allow the account holder to manually add keys

This change does NOT provide any means to remove keys once they've been
added.

Reviewed-by: Dan Bode