(#8583) Add lockout feature to accounts module
The previous behavior did not allow the end user to easily lock or
unlock an account.  This change adds a simple "locked" attribute in the
data file and if true will set the account to an invalid shell
appropriate for the platform Puppet is running on.

The SSH keys remain, but attempting to log in yields:

    ssh villain@centos56
    This account is currently not available.
    Connection to 172.16.214.129 closed.

Reviewed-by: Dan Bode