(#8583) Fix invalid shell on solaris
The lock behavior introduced in #8583 does not work correctly without
this patch.  When Puppet runs useradd, an error that /dev/null is not a
valid shell is raised.

With this patch, the behavior is that /usr/bin/false is accepted as a
valid shell and the account is created in a locked state.

Reviewed-by: Dan Bode