(#8579) Add example of system account management
This change simply adds an example of how to manage a service account.
It uses the RedHat reserved UID and GID of 52 to manage the puppet user
and group with a home directory of /var/lib/puppet.

The account is locked, which may not be desirable in the long run but is
acceptable for the initial release.