(#8678) Add relation edge from ~/.ssh to authorized_keys
Without this patch the ssh_authorized_key resources were not related to
the parent directory of ~/.ssh/ causing non-deterministic failures on
some but not all Puppet runs.

This patch adds an explicit relationship between the ssh_authorized_key
resources and their parent directory.

In addition, a variable declaration moves to the top of the block for
readability.

Reviewed-by: Dan Bode