(#8678) Always manage authorized_keys file permissions
Without this patch the accounts module was not managing the owner,
group, and mode of the authorized_keys file.  If Puppet was creating the
file during an enforcement run the permissions would be incorrectly set
to the EUID, EGID, and umask of the puppet process.

This patch always manages the owner, group and mode parameters of the
authorized_keys files.

In addition, make sure the authorized_keys file resource is managed
after all ssh_authorized_key resources.

Reviewed-by: Dan Bode