(#8678) Fix accounts::user user hash default lookup
Before this patch the accounts::user defined type was directly
referencing the namespace where the end user was expected to declare
default parameters for the user resource.

This is a problem because each resource declaration must be kept in step
with the main accounts class if the end user decided to customize this
location.

This change makes sure the main accounts class copies the default
parameters into it's own namespace.  It changes the behaviour of the
accounts::user defined type to default to looking up default settings in
the accounts namespace rather than accounts::data namespace.  This
resolves the problem by providing a consistent place to reference
user-customized data and provide just one place for the user to express
their customizations.

I also call out that the $data_namespace (defaults to 'accounts') used
by the accounts::user defined type should be declared in the catalog
before accounts::user resources are declared.  This resolves feedback
from Dan.

Closes: #8678

Reviewed-by: Dan Bode