(#8678) Refactor accounts::user removing hash use
Before this patch the accounts::user defined resource type required a
Hash data structure to be passed to the user_params parameter.  This
provided a clean interface to create_resources() but is not at all
friendly to the end user of Puppet.

In order to meet the goal of reducing the new user learning on-ramp and
to support the use case of resource defaults, this change removes the
requirement of the user_params hash and replaces the user-facing
interface with a thin proxy to all of the User resource types we're
managing.

Previous behavior:

    accounts::user { 'jeff':
      user_params => {
        ensure => present,
        uid    => 501,
        gid    => 501,
      }
      locked => false,
    }

New behavior with the patch:

    # Support Resource Defaults
    Accounts::User { shell => '/bin/zsh' }
    # Hide the complexity of create_resources()
    accounts::user { 'jeff':
      ensure => present,
      uid    => 501,
      gid    => 501,
      locked => false,
    }

Reviewed-by: Dan Bode