(#8678) Remove user_params hash from example data
The previous patch in this series removed the ability to supply the
nested user_params hash to the accounts::user defined resource type.

This patch in this series updates all of the examples, tests and sample
data included in the module to reflect this change.

It effectively moves the keys of the nested user_params hash "up one
level" in the data structures.

Reviewed-by: Dan Bode