(#8678) Manage no users and groups by default
Before this patch there was example data in the accounts::data namespace
in the $users_hash and $groups_has variables.  This data caused the
accounts module to manage users and groups if a simple manifest of node
default { include accounts } } is used.

This patch initializes $accounts::data::{user,group}_hash to an empty
Hash.  The accounts module does nothing by default as a result.

Reviewed-by: Dan Bode