(#8678) Get data from Puppet DSL namespace by default
Without this patch the accounts module has the default behavior of
trying to load data from a YAML data file.  This behavior requires
additional steps for the user to carry out before the module is usable.
Specifically, the user most copy example YAML files into
/etc/puppet/data, creating the directory if it does not exist.

This patch makes the namespace accounts::data the default location for
data.  The module ships with usable sane default data in this namespace.
This patch makes the module easier to consume by reducing the on-ramp to
a simple puppet-module install accounts command, and then:

    node default {
      accounts::user { 'dan': }
      accounts::user { 'jeff': }
    }

Reviewed-by: Dan Bode