(#8679) Mention accounts::user first in the README
Without this patch, the README mentions the accounts::user defined
resource type last in the Declaring User Accounts section.

Based on feedback from Nigel, we think the accounts::user defined
resource type is likely the first things end-users will touch.  As a
result, it should be mentioned first in the section.

This patch updates the README to call out the accounts::user defined
resource type early, provide a quick-start, and provide hints on where
to look for more advanced examples.

Reviewed-by: Dan Bode