(#8678) Remove use of create_resources hash in accounts::user
Without this patch the accounts::user defined resource type constructs a
Hash and passes the Hash to create_resources().  This isn't required and
is unclear to the end user why we may be doing this.  It's effectively
an artifact from the previous design and implementation.

This patch switches to the puppety style of declaring the user resource
directly.

In addition, the validation and munging of the resource parameters has
been cleaned up.  _real variables have been removed where no munging is
being done.

Reviewed-by: Dan Bode