(#8794) Rename whole_line to file_line in accounts class
Before this patch, the accounts module will not work with stdlib-2.0.0
because of the API breaking change to {append,whole}_line.

As of stdlib-2.0.0 the API is to use file_line.

This patch updates the accounts class to use the file_line resource.

In addition, a suite of unit tests have been added to test all of the
class parameters to the accounts class.  These tests will begin to fail
if the stdlib API changes again for file_line.