(#8578) Do not manage sudoers by default
Without this change the accounts class would automatically manage lines
in the sudoers file when included into the catalog.

This behavior is undesirable because the end user might expect the
accounts class to only manage accounts when declared.

This patch makes the management of the sudoers file off by default.  The
end user must explicitly enable the behavior using the manage_sudoers
class parameter like so:

    node default {
      class accounts {
        manage_sudoers => true,
      }
    }

In addition this patch adds test coverage for this expectation and the
expectation sudoers is not managed on Solaris even when explicitly
declared.