(#8718) Add test coverage for anchor resources
Without this patch the accounts module did not contain any test coverage
for the anchor resources.  This could result in refactors of the module
not catching resource containment problems.

This patch simply verifies the anchor resources are present in the
catalog when the accounts class is present in the catalog.

The comment for the Solaris test has also been repositioned slightly
because it was in the wrong location.