(#8829) Update FIXME tags in module
Before this patch some FIXME tags were out of date relative to the
current state of the accounts module.

This patch updates the FIXME tag related to create_resources() and 2.7.
In addition, the patch removes the FIXME tag regarding the $users_hash
scope issue which is no longer an issue and is within scope by design.