(PE-4130) Loosening dependency on puppetlabs-stdlib.
Prior to this commit, we had a firm dependency on stdlib 3.2.x.  This, unfortunately,
kept users from being able to upgrade stdlib, and from relying on modules others had
written that depended on a more recent stdlib.  This change loosens the explicit
version range we express a dependency on, and combined with a new supported version of
stdlib in the 4.x series, will give users the ability to opt into upgrading their
installations.