(FM-3217) Manage solaris 11 home dirs
Solaris 11 home dirs are mounted at /export/home and are zfs mount
points. This module cannot predict the zpool on which to create the
homedirectory and so leaves it up to `useradd -m`. The zfs mount may
also be pre-created before the accounts resource is run.

Additionally, all resources in the users home directory will not be
managed if managehome is false, but will always be managed if it is true
(ie, no more replace => false)