fix problematic parsing of keyspec
The original had some flaws. The `.*` statement is an eager match and could eat spaces. An example:

`options ssh-xyz key name with spaces` would be matched as:

```
[
  "options ",
  "options",
  "ssh-xyz key name",
  "ssh",
  "with",
  "spaces"
]
```

The new regex would match it like this:
```
[
  "options ",
  "options",
  "ssh-xyz",
  "ssh",
  "key",
  "name with spaces"
]
```

Another example: `ssh-xyz key ssh name with spaces`
```
[
  "ssh-xyz key ",
  "ssh-xyz key",
  "ssh name",
  "ssh",
  "with",
  "spaces"
]
```
vs
```
[
  "",
  "",
  "ssh-xyz",
  "ssh",
  "key",
  "ssh name with spaces"
]
```

Workaround until a fix is released: carefully manage the spaces and do not use `ssh` or `ecdsa-sha2` in unexpected places (certainly no spaces in the name).

Thanks to @ekoster for finding this.