More lenient username checks (allow capitals)
Since there is an exception for periods on the basis that "some installations allow it", i don't see why there shouldn't be an exception for capitals. 

Seperately, older versions of this puppet module had no restriction on capitals, so enforcing them now breaks things even though the underlying OS has no problem with it, this does not fare well with users who potentially need to rework puppet code & fix or replace existing systems.